fix:updatd

+++ src/main/java/com/nhnacademy/bookstore/purchase/bookCart/service/impl/BookCartGuestServiceImpl.java
@@ -13,10 +13,12 @@ import com.nhnacademy.bookstore.purchase.bookCart.service.BookCartGuestService;
 import com.nhnacademy.bookstore.purchase.cart.exception.CartDoesNotExistException;
 import com.nhnacademy.bookstore.purchase.cart.repository.CartRepository;
 import java.time.ZonedDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,6 +30,7 @@ import lombok.RequiredArgsConstructor;
  *
  * @author 김병우
  */
+@Slf4j
 @Transactional
 @Service
 @RequiredArgsConstructor
@@ -41,42 +44,37 @@ public class BookCartGuestServiceImpl implements BookCartGuestService {
      * 도서장바구니 추가.
      *
      * @param bookId 도서아이디
-     * @param cartId 장바구니아이디
      * @param quantity 수량
      * @return bookCartId
      */
     @Override
-    public Long createBookCart(Long bookId, Long cartId, int quantity) {
+    public Long createBookCart(Long bookId, int quantity) {
         Book book = bookRepository.findById(bookId)
                 .orElseThrow(() -> new BookDoesNotExistException("도서가 존재하지 않습니다"));
 
-        if (Objects.isNull(cartId)) {
-            Cart newCart = new Cart();
-            cartRepository.save(newCart);
-            cartId = newCart.getId();
-        }
+        Cart cart = new Cart();
+        cartRepository.save(cart);
+        long cartId = cart.getId();
 
-        Cart cart = cartRepository.findById(cartId)
-                .orElseThrow(() -> new CartDoesNotExistException("카트가 존재하지 않습니다"));
-
-        hasDataToLoad(cart.getId());
+        //hasDataToLoad(cart.getId());
 
         BookCart bookCart = new BookCart(quantity, ZonedDateTime.now(), book, cart);
 
         bookCartRepository.save(bookCart);
 
-        bookCartRedisRepository.create(
-                cartId.toString(),
-                bookCart.getId(),
-                ReadBookCartGuestResponse.builder()
-                        .bookCartId(bookCart.getId())
-                        .price(book.getPrice())
-                        .title(book.getTitle())
-                        .quantity(bookCart.getQuantity())
-                        .build()
-        );
+//        bookCartRedisRepository.create(
+//                Long.toString(cartId),
+//                bookCart.getId(),
+//                ReadBookCartGuestResponse.builder()
+//                        .bookCartId(bookCart.getId())
+//                        .price(book.getPrice())
+//                        .url(book.getBookImageList().getFirst().getUrl())
+//                        .title(book.getTitle())
+//                        .quantity(bookCart.getQuantity())
+//                        .build()
+//        );
 
-        return bookCart.getId();
+        return cartId;
     }
 
     /**
@@ -95,22 +93,29 @@ public class BookCartGuestServiceImpl implements BookCartGuestService {
         Cart cart = cartRepository.findById(cartId)
                 .orElseThrow(() -> new CartDoesNotExistException(cartId + "가 존재하지 않습니다"));
 
-        BookCart bookCart = bookCartRepository.findByBookAndCart(book, cart)
-                .orElseThrow(() -> new BookCartDoesNotExistException("북카트 존재하지 않습니다."));
 
-        hasDataToLoad(cartId);
+        Optional<BookCart> optionalBookCart = bookCartRepository.findByBookAndCart(book, cart);
+        if(optionalBookCart.isEmpty()){
+            bookCartRepository.save(new BookCart(quantity,book,cart));
+            return cartId;
+        }
+
+        //hasDataToLoad(cartId);
+
+        BookCart bookCart = optionalBookCart.get();
 
         int amount = bookCart.getQuantity() + quantity;
         if (amount > 0) {
             bookCart.setQuantity(amount);
+
             bookCartRepository.save(bookCart);
-            bookCartRedisRepository.update(cartId.toString(), bookCart.getId(), amount);
+            //bookCartRedisRepository.update(cartId.toString(), bookCart.getId(), amount);
         } else {
             bookCartRepository.delete(bookCart);
-            bookCartRedisRepository.delete(cartId.toString(), bookCart.getId());
+            //bookCartRedisRepository.delete(cartId.toString(), bookCart.getId());
         }
 
-        return bookCart.getId();
+        return cart.getId();
     }
 
     @Override
@@ -119,9 +124,9 @@ public class BookCartGuestServiceImpl implements BookCartGuestService {
                 .orElseThrow(() -> new BookCartDoesNotExistException("북카트 존재하지 않습니다."));
 
         bookCartRepository.delete(bookCart);
-        bookCartRedisRepository.delete(cartId.toString(), bookCart.getId());
+        //bookCartRedisRepository.delete(cartId.toString(), bookCart.getId());
 
-        return bookCart.getId();
+        return cartId;
     }
 
 
@@ -134,13 +139,13 @@ public class BookCartGuestServiceImpl implements BookCartGuestService {
      */
     @Override
     public List<ReadBookCartGuestResponse> readAllBookCart(Long cartId) {
-        List<ReadBookCartGuestResponse> bookCartGuestResponseList = bookCartRedisRepository.readAllHashName(cartId.toString());
-        if (!bookCartGuestResponseList.isEmpty()) {
-            return  bookCartGuestResponseList;
-        }
-
-        return  hasDataToLoad(cartId);
-
+//        List<ReadBookCartGuestResponse> bookCartGuestResponseList =  bookCartRedisRepository.readAllHashName(cartId.toString());
+//        if (!bookCartGuestResponseList.isEmpty()) {
+//            return  bookCartGuestResponseList;
+//        }
+//
+//        return  hasDataToLoad(cartId);
+        return readAllFromDb(cartId);
     }
 
     /**
@@ -164,16 +169,20 @@ public class BookCartGuestServiceImpl implements BookCartGuestService {
      * @return DB 도서 장바구니 목록
      */
     private List<ReadBookCartGuestResponse> readAllFromDb(Long cartId) {
-        Optional<Cart> cart = cartRepository.findById(cartId);
+        List<BookCart> list = bookCartRepository.findAllByCartId(cartId);
+        List<ReadBookCartGuestResponse> listDto = new ArrayList<>();
+        log.info("{}", list);
+        for (BookCart bookCart : list) {
+            listDto.add(ReadBookCartGuestResponse.builder()
+                    .bookCartId(bookCart.getId())
+                    .bookId(bookCart.getBook().getId())
+                    .price(bookCart.getBook().getPrice())
+                    .url(bookCart.getBook().getBookImageList().getFirst().getUrl())
+                    .title(bookCart.getBook().getTitle())
+                    .quantity(bookCart.getQuantity())
+                    .build());
+        }
 
-        return cart.map(value -> bookCartRepository.findAllByCart(value)
-                .stream()
-                .map(bookCart -> ReadBookCartGuestResponse.builder()
-                        .bookCartId(bookCart.getId())
-                        .price(bookCart.getBook().getPrice())
-                        .title(bookCart.getBook().getTitle())
-                        .quantity(bookCart.getQuantity())
-                        .build())
-                .toList()).orElseGet(List::of);
+        return listDto;
     }
 }
